+++ CHANGELOG.md
@@ -10,9 +10,9 @@ For a full diff see [`0.11.0...master`](https://github.com/localheinz/phpstan-ru
 
 ### Added
 
-* Added `Methods\NoParameterWithContainerTypeDeclarationRule`, which reports an error when a method has a type declaration that corresponds to a known dependency injection container or service locator ([#122](https://github.com/localheinz/phpstan-rules/pull/122)), by [@dmecke](https://github.com/dmecke)
+* Added `Methods\NoParameterWithContainerTypeDeclarationRule`, which reports an error when a method has a type declaration that corresponds to a known dependency injection container or service locator ([#122](https://github.com/localheinz/phpstan-rules/pull/122)), by [@localheinz](https://github.com/localheinz)
 
-## [`0.11.0`](https://github.com/localheinz/phpstan-rules/releases/tag/0.10.0)
+## [`0.11.0`](https://github.com/localheinz/phpstan-rules/releases/tag/0.11.0)
 
 For a full diff see [`0.10.0...0.11.0`](https://github.com/localheinz/phpstan-rules/compare/0.10.0...0.11.0).